% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprod.r
\name{crossprod_game}
\alias{crossprod_game}
\title{crossprod_game}
\usage{
crossprod_game(x)
}
\arguments{
\item{x}{A shaq.}
}
\value{
A regular matrix.
}
\description{
Crossproduct on a gpu.
}
\section{Communication}{

The operation consists of a local crossproduct, followed by an
\code{allreduce()} call, quadratic on the number of columns.

For \code{crossprod()}, if the matrix distribution is poorly balanced
(specifically, if any rank has fewer rows than columns), then an inefficient
method is used. Similarly for \code{tcrossprod()} if the number of local rows
is greater than the number of local columns.
}

