% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimrgame-package.r
\docType{package}
\name{dimrgame-package}
\alias{dimrgame-package}
\title{dimrgame}
\description{
Clustering analysis methods for GPUs.
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\keyword{Package}
