% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.r
\name{svd_game}
\alias{svd_game}
\title{svd_game}
\usage{
svd_game(x)
}
\arguments{
\item{x}{A shaq.}
}
\value{
A named list, as in kazaam (which mimics R's \code{svd()}).
}
\description{
Singular value decomposition on a gpu.
}
